Changed Dockerfile and CICO script to enable easier manual Docker builds
Motivation
----------
Right now the Dockerfile and CICO script take the WAR file from `/target`
which is not where it's built, which is `/web/target`. Which means that
for manual Docker builds you first have to copy the WAR. THat's an
unnecessary extra step.

Modifications
-------------
The paths have simply been changed to `/web/target`.

Result
------
Same behaviour as before except that manual Docker builds are now easier.